+++ src/hanami/src/api/http/dataset/mnist/finalize_mnist_dataset_v1_0.cpp
@@ -179,12 +179,12 @@ FinalizeMnistDataSetV1M0::convertMnistData(const std::string& filePath,
     // create description
     json description;
     json inputDescrEntry;
-    inputDescrEntry["start_column"] = 0;
-    inputDescrEntry["end_column"] = pictureSize;
+    inputDescrEntry["column_start"] = 0;
+    inputDescrEntry["column_end"] = pictureSize;
     description["picture"] = inputDescrEntry;
     json outputDescrEntry;
-    outputDescrEntry["start_column"] = pictureSize;
-    outputDescrEntry["end_column"] = pictureSize + 10;
+    outputDescrEntry["column_start"] = pictureSize;
+    outputDescrEntry["column_end"] = pictureSize + 10;
     description["label"] = outputDescrEntry;
 
     // initialize file